submit 3rd laboratory exercise

+++ backend/src/main/java/hr/fer/zari/or/backend/repository/ArtistRepository.java
@@ -7,5 +7,5 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface ArtistRepository extends JpaRepository<Artist, Integer> {
+public interface ArtistRepository extends JpaRepository<Artist, Long> {
 }
